refactor(adapters): extract shared memory scoring helpers to _memory_scoring
- Move _recency_score, _format_episode_block, _CHARS_PER_TOKEN,
  _AVG_EPISODE_CHARS, _OUTCOME_WEIGHTS to src/ravn/adapters/_memory_scoring.py
- Extract prefetch budget-truncation logic into build_prefetch_context()
- Extract session-grouping logic into build_session_summaries()
- Import shared helpers in both SqliteMemoryAdapter and PostgresMemoryAdapter
- Use CTEs in postgres_memory query_episodes and search_sessions to avoid
  computing websearch_to_tsquery twice per query
