+++ src/ravn/adapters/postgres_memory.py
@@ -12,47 +12,26 @@ scores are comparable across backends.
 from __future__ import annotations
 
 import json
-import math
 import os
-from collections import defaultdict
 from datetime import UTC, datetime
 from typing import Any
 
 import asyncpg
 
+from ravn.adapters._memory_scoring import (
+    _AVG_EPISODE_CHARS,
+    _CHARS_PER_TOKEN,
+    _OUTCOME_WEIGHTS,
+    _recency_score,
+    build_prefetch_context,
+    build_session_summaries,
+)
 from ravn.domain.models import Episode, EpisodeMatch, Outcome, SessionSummary, SharedContext
 from ravn.ports.memory import MemoryPort
 
 # ---------------------------------------------------------------------------
-# Constants
-# ---------------------------------------------------------------------------
-
-# Approximate chars per token for budget estimation.
-_CHARS_PER_TOKEN = 4
-
-# Estimated average character length of a single episode row, used to compute
-# the FTS LIMIT when searching for sessions.
-_AVG_EPISODE_CHARS = 200
-
-# Outcome weights for combined scoring.
-_OUTCOME_WEIGHTS: dict[str, float] = {
-    Outcome.SUCCESS: 1.0,
-    Outcome.PARTIAL: 0.7,
-    Outcome.FAILURE: 0.3,
-}
-
-
-# ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
-
-
-def _recency_score(timestamp: datetime, half_life_days: float) -> float:
-    """Compute exponential decay recency in [0, 1] given episode age."""
-    now = datetime.now(UTC)
-    ts = timestamp if timestamp.tzinfo is not None else timestamp.replace(tzinfo=UTC)
-    age_days = (now - ts).total_seconds() / 86400.0
-    return math.exp(-age_days * math.log(2) / half_life_days)
 
 
 def _combined_score(
@@ -65,19 +44,6 @@ def _combined_score(
     recency = _recency_score(timestamp, half_life_days)
     weight = _OUTCOME_WEIGHTS.get(outcome, 0.5)
     return ts_rank * recency * weight
-
-
-def _format_episode_block(episode: Episode) -> str:
-    """Format a single episode for injection into the system prompt."""
-    ts = episode.timestamp.strftime("%Y-%m-%d")
-    outcome = episode.outcome.upper()
-    tags_str = ", ".join(episode.tags) if episode.tags else "general"
-    tools_str = ", ".join(episode.tools_used) if episode.tools_used else "none"
-    return (
-        f"[{ts}] [{outcome}] {episode.task_description}\n"
-        f"Tags: {tags_str} | Tools: {tools_str}\n"
-        f"{episode.summary}"
-    )
 
 
 def _row_to_episode(row: asyncpg.Record | dict[str, Any]) -> Episode:
@@ -245,12 +211,13 @@ class PostgresMemoryAdapter(MemoryPort):
         async with pool.acquire() as conn:
             rows = await conn.fetch(
                 """
+                WITH q AS (SELECT websearch_to_tsquery('english', $1) AS tsq)
                 SELECT
                     episode_id, session_id, timestamp, summary,
                     task_description, tools_used, outcome, tags, embedding,
-                    ts_rank(search_vector, websearch_to_tsquery('english', $1)) AS rank_score
-                FROM ravn_episodes
-                WHERE search_vector @@ websearch_to_tsquery('english', $1)
+                    ts_rank(search_vector, q.tsq) AS rank_score
+                FROM ravn_episodes, q
+                WHERE search_vector @@ q.tsq
                 ORDER BY rank_score DESC
                 LIMIT $2
                 """,
@@ -285,23 +252,8 @@ class PostgresMemoryAdapter(MemoryPort):
         )
         if not matches:
             return ""
-
         budget_chars = self._prefetch_budget * _CHARS_PER_TOKEN
-        blocks: list[str] = []
-        used = 0
-        for match in matches:
-            block = _format_episode_block(match.episode)
-            if used + len(block) > budget_chars:
-                break
-            blocks.append(block)
-            used += len(block) + 1  # +1 for separator
-
-        if not blocks:
-            return ""
-
-        separator = "\n\n---\n\n"
-        body = separator.join(blocks)
-        return f"## Relevant Past Context\n\n{body}"
+        return build_prefetch_context(matches, budget_chars)
 
     async def search_sessions(
         self,
@@ -316,12 +268,13 @@ class PostgresMemoryAdapter(MemoryPort):
         async with pool.acquire() as conn:
             rows = await conn.fetch(
                 """
+                WITH q AS (SELECT websearch_to_tsquery('english', $1) AS tsq)
                 SELECT
                     episode_id, session_id, timestamp, summary,
                     task_description, tools_used, outcome, tags, embedding
-                FROM ravn_episodes
-                WHERE search_vector @@ websearch_to_tsquery('english', $1)
-                ORDER BY ts_rank(search_vector, websearch_to_tsquery('english', $1)) DESC
+                FROM ravn_episodes, q
+                WHERE search_vector @@ q.tsq
+                ORDER BY ts_rank(search_vector, q.tsq) DESC
                 LIMIT $2
                 """,
                 query,
@@ -331,42 +284,8 @@ class PostgresMemoryAdapter(MemoryPort):
         if not rows:
             return []
 
-        session_episodes: dict[str, list[Episode]] = defaultdict(list)
-        for row in rows:
-            ep = _row_to_episode(row)
-            session_episodes[ep.session_id].append(ep)
-
-        summaries: list[SessionSummary] = []
-        for session_id, episodes in session_episodes.items():
-            episodes.sort(key=lambda e: e.timestamp)
-            last_active = max(e.timestamp for e in episodes)
-            all_tags: list[str] = []
-            for ep in episodes:
-                all_tags.extend(ep.tags)
-            unique_tags = list(dict.fromkeys(all_tags))
-
-            lines: list[str] = []
-            total_chars = 0
-            for ep in episodes:
-                line = f"- [{ep.outcome.upper()}] {ep.task_description}: {ep.summary}"
-                if total_chars + len(line) > self._session_search_truncate_chars:
-                    break
-                lines.append(line)
-                total_chars += len(line)
-
-            summary_text = "\n".join(lines)
-            summaries.append(
-                SessionSummary(
-                    session_id=session_id,
-                    summary=summary_text,
-                    episode_count=len(episodes),
-                    last_active=last_active,
-                    tags=unique_tags[:10],
-                )
-            )
-
-        summaries.sort(key=lambda s: s.last_active, reverse=True)
-        return summaries[:limit]
+        episodes = [_row_to_episode(row) for row in rows]
+        return build_session_summaries(episodes, limit, self._session_search_truncate_chars)
 
     def inject_shared_context(self, context: SharedContext) -> None:
         self._shared_context = context
@@ -381,16 +300,12 @@ class PostgresMemoryAdapter(MemoryPort):
     def _require_pool(self) -> asyncpg.Pool:
         """Return the pool, raising RuntimeError if not initialized."""
         if self._pool is None:
-            raise RuntimeError(
-                "PostgresMemoryAdapter not initialized. Call initialize() first."
-            )
+            raise RuntimeError("PostgresMemoryAdapter not initialized. Call initialize() first.")
         return self._pool
 
     async def _detect_pgvector(self) -> bool:
         """Return True if the pgvector extension is installed in this database."""
         pool = self._require_pool()
         async with pool.acquire() as conn:
-            row = await conn.fetchrow(
-                "SELECT 1 FROM pg_extension WHERE extname = 'vector'"
-            )
+            row = await conn.fetchrow("SELECT 1 FROM pg_extension WHERE extname = 'vector'")
         return row is not None
